+++ VS2019/audio_analysis_lib.vcxproj
@@ -21,7 +21,6 @@
   <ItemGroup>
     <ClCompile Include="..\src\impl\fft_component.cpp" />
     <ClCompile Include="..\src\impl\input_capture.cpp" />
-    <ClCompile Include="..\src\impl\input_capture_impl.cpp" />
     <ClCompile Include="..\src\impl\window_functions.cpp" />
     <ClCompile Include="..\src\impl\world_cheaptrick_sp.cpp" />
     <ClCompile Include="..\src\impl\world_d4c_ap.cpp" />
@@ -32,7 +31,6 @@
     <ClInclude Include="..\src\define_macro.hpp" />
     <ClInclude Include="..\src\fft_component.hpp" />
     <ClInclude Include="..\src\fft_exception.hpp" />
-    <ClInclude Include="..\src\impl\input_capture_impl.hpp" />
     <ClInclude Include="..\src\input_capture.hpp" />
     <ClInclude Include="..\src\lib_audio_analysis.h" />
     <ClInclude Include="..\src\window_functions.hpp" />
