+++ VS2019/audio_analysis_lib.vcxproj.filters
@@ -21,9 +21,6 @@
     <ClCompile Include="..\src\impl\input_capture.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
-    <ClCompile Include="..\src\impl\input_capture_impl.cpp">
-      <Filter>ソース ファイル</Filter>
-    </ClCompile>
     <ClCompile Include="..\src\impl\window_functions.cpp">
       <Filter>ソース ファイル</Filter>
     </ClCompile>
@@ -41,9 +38,6 @@
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
-    <ClInclude Include="..\src\impl\input_capture_impl.hpp">
-      <Filter>ヘッダー ファイル</Filter>
-    </ClInclude>
     <ClInclude Include="..\src\define_macro.hpp">
       <Filter>ヘッダー ファイル</Filter>
     </ClInclude>
